+++ livekit-api/src/lib.rs
@@ -40,7 +40,8 @@ pub mod signal_client;
 mod http_client;
 
 // Region-discovery helpers shared by the signaling region provider
-// (signal_client::region) and the API failover region cache (services::failover).
+// (signal_client::region_url_provider) and the API failover region cache
+// (services::failover).
 #[cfg(any(
     feature = "signal-client-tokio",
     feature = "signal-client-async",
